+++ backend/core/src/jurisdictions/entities/jurisdiction.entity.ts
@@ -15,7 +15,6 @@ import { ValidationsGroupsEnum } from "../../shared/types/validations-groups-enu
 import { Language } from "../../shared/types/language-enum"
 import { Expose, Type } from "class-transformer"
 import { Preference } from "../../preferences/entities/preference.entity"
-import { User } from "../../auth/entities/user.entity"
 
 @Entity({ name: "jurisdictions" })
 export class Jurisdiction extends AbstractEntity {
@@ -84,4 +83,10 @@ export class Jurisdiction extends AbstractEntity {
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })
   @IsBoolean({ groups: [ValidationsGroupsEnum.default] })
   enableAccessibilityFeatures: boolean | null
+
+  @Column({ type: "boolean", nullable: false })
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsBoolean({ groups: [ValidationsGroupsEnum.default] })
+  enableUtilitiesIncluded: boolean | null
 }