+++ backend/core/src/listings/listings.module.ts
@@ -11,7 +11,8 @@ import { TranslationsModule } from "../translations/translations.module"
 import { AmiChart } from "../ami-charts/entities/ami-chart.entity"
 import { ListingFeatures } from "./entities/listing-features.entity"
 import { ActivityLogModule } from "../activity-log/activity-log.module"
-import { ListingRepository } from "./repositories/listing.repository"
+import { ListingRepository } from "./db/listing.repository"
+import { ListingUtilities } from "./entities/listing-utilities.entity"
 
 @Module({
   imports: [
@@ -23,6 +24,7 @@ import { ListingRepository } from "./repositories/listing.repository"
       AmiChart,
       ListingRepository,
       ListingFeatures,
+      ListingUtilities,
     ]),
     AuthModule,
     TranslationsModule,