+++ backend/core/types/src/backend-swagger.ts
@@ -3830,6 +3830,9 @@ export interface Jurisdiction {
 
   /**  */
   enableAccessibilityFeatures: boolean
+
+  /**  */
+  enableUtilitiesIncluded: boolean
 }
 
 export interface User {
@@ -4269,6 +4272,9 @@ export interface JurisdictionCreate {
   enableAccessibilityFeatures: boolean
 
   /**  */
+  enableUtilitiesIncluded: boolean
+
+  /**  */
   programs: Id[]
 
   /**  */
@@ -4311,6 +4317,9 @@ export interface JurisdictionUpdate {
 
   /**  */
   enableAccessibilityFeatures: boolean
+
+  /**  */
+  enableUtilitiesIncluded: boolean
 
   /**  */
   programs: Id[]
@@ -4591,6 +4600,9 @@ export interface Preference {
   links?: PreferenceLink[]
 
   /**  */
+  jurisdictions?: IdName[]
+
+  /**  */
   id: string
 
   /**  */
@@ -4864,6 +4876,32 @@ export interface ListingFeatures {
   mobility?: boolean
 }
 
+export interface ListingUtilities {
+  /**  */
+  water?: boolean
+
+  /**  */
+  gas?: boolean
+
+  /**  */
+  trash?: boolean
+
+  /**  */
+  sewer?: boolean
+
+  /**  */
+  electricity?: boolean
+
+  /**  */
+  cable?: boolean
+
+  /**  */
+  phone?: boolean
+
+  /**  */
+  internet?: boolean
+}
+
 export interface Listing {
   /**  */
   referralApplication?: ApplicationMethod
@@ -4990,6 +5028,9 @@ export interface Listing {
 
   /**  */
   features?: ListingFeatures
+
+  /**  */
+  utilities?: ListingUtilities
 
   /**  */
   id: string
@@ -5551,6 +5592,9 @@ export interface ListingCreate {
 
   /**  */
   features?: ListingFeatures
+
+  /**  */
+  utilities?: ListingUtilities
 }
 
 export interface ListingEventUpdate {
@@ -5975,6 +6019,9 @@ export interface ListingUpdate {
 
   /**  */
   features?: ListingFeatures
+
+  /**  */
+  utilities?: ListingUtilities
 }
 
 export interface PreferencesFilterParams {
@@ -6000,6 +6047,9 @@ export interface PreferenceCreate {
 
   /**  */
   formMetadata?: FormMetadata
+
+  /**  */
+  jurisdictions?: IdName[]
 }
 
 export interface PreferenceUpdate {
@@ -6017,6 +6067,9 @@ export interface PreferenceUpdate {
 
   /**  */
   formMetadata?: FormMetadata
+
+  /**  */
+  jurisdictions?: IdName[]
 
   /**  */
   id: string