+++ sites/partners/src/listings/PaperListingForm/sections/AdditionalFees.tsx
@@ -1,13 +1,34 @@
-import React from "react"
+import React, { useContext, useMemo } from "react"
 import { useFormContext } from "react-hook-form"
-import { t, GridSection, Field, Textarea } from "@bloom-housing/ui-components"
+import { t, GridSection, Field, Textarea, FieldGroup, ViewItem } from "@bloom-housing/ui-components"
 import { fieldHasError, fieldMessage } from "../../../../lib/helpers"
+import { AuthContext, listingUtilities } from "@bloom-housing/shared-helpers"
+import { ListingUtilities } from "@bloom-housing/backend-core/types"
 
-const AdditionalFees = () => {
+type AdditionalFeesProps = {
+  existingUtilities: ListingUtilities
+}
+
+const AdditionalFees = (props: AdditionalFeesProps) => {
   const formMethods = useFormContext()
-
+  const { profile } = useContext(AuthContext)
   // eslint-disable-next-line @typescript-eslint/unbound-method
-  const { register, errors, clearErrors } = formMethods
+  const { register, watch, errors, clearErrors } = formMethods
+
+  const jurisdiction = watch("jurisdiction.id")
+
+  const utilitiesFields = useMemo(() => {
+    return listingUtilities.map((utility) => ({
+      id: utility,
+      label: t(`listings.utilities.${utility}`),
+      defaultChecked: props.existingUtilities ? props.existingUtilities[utility] : false,
+      register,
+    }))
+  }, [props.existingUtilities, register])
+
+  const enableUtilitiesIncluded = profile?.jurisdictions?.find((j) => j.id === jurisdiction)
+    ?.enableUtilitiesIncluded
+
   return (
     <div>
       <GridSection
@@ -73,6 +94,19 @@ const AdditionalFees = () => {
             register={register}
           />
         </GridSection>
+        {enableUtilitiesIncluded && (
+          <GridSection columns={1}>
+            <ViewItem label={t("listings.sections.utilities")}>
+              <FieldGroup
+                type="checkbox"
+                name="listingUtilities"
+                fields={utilitiesFields}
+                register={register}
+                fieldGroupClassName="grid grid-cols-2 mt-4"
+              />
+            </ViewItem>
+          </GridSection>
+        )}
       </GridSection>
     </div>
   )