+++ ui-components/src/page_components/listing/listing_sidebar/ExpandableSection.tsx
@@ -3,7 +3,7 @@ import Markdown from "markdown-to-jsx"
 import { ExpandableContent } from "../../../actions/ExpandableContent"
 
 export interface ExpandableSectionProps {
-  content: string
+  content: string | React.ReactNode
   expandableContent?: string
   strings: {
     title: string
@@ -18,7 +18,11 @@ const ExpandableSection = ({ content, expandableContent, strings }: ExpandableSe
     <section className="aside-block">
       <h4 className="text-caps-underline">{strings.title}</h4>
       <div className="text-tiny text-gray-750">
-        <Markdown options={{ disableParsingRawHTML: false }}>{content}</Markdown>
+        {typeof content === "string" ? (
+          <Markdown options={{ disableParsingRawHTML: false }}>{content}</Markdown>
+        ) : (
+          content
+        )}
         {expandableContent && (
           <div className={"mt-2"}>
             <ExpandableContent strings={{ readMore: strings.readMore, readLess: strings.readLess }}>