+++ ui-components/src/page_components/listing/listing_sidebar/SidebarBlock.tsx
@@ -1,0 +1,23 @@
+import * as React from "react"
+import { Heading, HeaderType, HeadingProps } from "../../../headers/Heading"
+
+export interface SidebarBlockProps extends Omit<HeadingProps, "children" | "className"> {
+  children: React.ReactNode
+  className?: string
+  title?: string
+}
+
+const SidebarBlock = ({ children, className, title, priority, style }: SidebarBlockProps) => {
+  return (
+    <section className={`aside-block ${className ? className : ""}`}>
+      {title && (
+        <Heading priority={priority ?? 4} style={style ?? "sidebarHeader"}>
+          {title}
+        </Heading>
+      )}
+      <div className="text-tiny text-gray-750">{children}</div>
+    </section>
+  )
+}
+
+export { SidebarBlock as default, SidebarBlock }