+++ sites/partners/CHANGELOG.md
@@ -3,6 +3,116 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.1-alpha.24](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.23...@bloom-housing/partners@5.0.1-alpha.24) (2022-07-13)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [5.0.1-alpha.23](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.22...@bloom-housing/partners@5.0.1-alpha.23) (2022-07-13)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [5.0.1-alpha.22](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.21...@bloom-housing/partners@5.0.1-alpha.22) (2022-07-12)
+
+
+### Features
+
+* modal styling updates ([#2857](https://github.com/bloom-housing/bloom/issues/2857)) ([4faf6de](https://github.com/bloom-housing/bloom/commit/4faf6de448a7fff8d4fa1a7a1144674ace311732))
+
+
+
+
+
+## [5.0.1-alpha.21](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.20...@bloom-housing/partners@5.0.1-alpha.21) (2022-07-12)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [5.0.1-alpha.20](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.19...@bloom-housing/partners@5.0.1-alpha.20) (2022-07-12)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [5.0.1-alpha.19](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.18...@bloom-housing/partners@5.0.1-alpha.19) (2022-07-12)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [5.0.1-alpha.18](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.17...@bloom-housing/partners@5.0.1-alpha.18) (2022-07-11)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [5.0.1-alpha.17](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.16...@bloom-housing/partners@5.0.1-alpha.17) (2022-07-11)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [5.0.1-alpha.16](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.15...@bloom-housing/partners@5.0.1-alpha.16) (2022-07-11)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [5.0.1-alpha.15](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.14...@bloom-housing/partners@5.0.1-alpha.15) (2022-07-08)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [5.0.1-alpha.14](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.13...@bloom-housing/partners@5.0.1-alpha.14) (2022-07-07)
+
+
+### Features
+
+* populate jurisdictional preferences on settings page ([#2828](https://github.com/bloom-housing/bloom/issues/2828)) ([2ea5646](https://github.com/bloom-housing/bloom/commit/2ea5646a49dc6782fe0e83e942e1917929fde153))
+
+
+
+
+
+## [5.0.1-alpha.13](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.12...@bloom-housing/partners@5.0.1-alpha.13) (2022-07-06)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [5.0.1-alpha.12](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.11...@bloom-housing/partners@5.0.1-alpha.12) (2022-07-06)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
 ## [5.0.1-alpha.11](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@5.0.1-alpha.10...@bloom-housing/partners@5.0.1-alpha.11) (2022-07-02)
 
 **Note:** Version bump only for package @bloom-housing/partners